카테고리, 활동, 기록 서버 연결 중

+++ app/src/main/java/com/example/buildupfrontend/retrofit/Request/CategoryRequest.kt
@@ -2,11 +2,11 @@ package com.example.buildupfrontend.retrofit.Request
 
 data class CategoryRequest(
     val categoryName: String,
-    val iconId: Long
+    val iconId: Int
 )
 
 data class EditCategoryRequest(
-    val id: Long,
+    val id: Int,
     val categoryName: String,
-    val iconId: Long
+    val iconId: Int
 )
